Edits for Package File Service
Missing quest objectives

The first section is called package, file, resource, but the first paragraph speaks about package, file, service. I would be sure to clearly identify the relationship or difference.

Line 171: For me, the relationship between 'source' and 'file' was unclear In particular, this: "With this source file in place, you can now use it to define the content of a`file` resource. The `file` resource takes a `source` parameter, which allows you to specify a source file that will define the content of the managed file." It was hard for me to follow the order of operations and what comes first--source or file?

Line 236: Should [Unit] be indented?