Forge Edits
Like 61 and 68: The word IMAGE is there but no image.
Line 75: Missing part of a sentence here, and I was not sure how to fix it.

There are no task headers throughout this quest as there are in other quests.
There is an active ToDo for you in here
There is no Review for this quest.

Since stuff is not yet finished, I will submit the PR but you can comment and let me know when you have added the missing content. I can come back in and review it again before committing.