Edits for defined resources_types
There are no quest tasks/headings in this quest.

Line 183: Since this is something the user has to do, I would give them an explicit command of what they will do next above this line.

I think that this quest, like the others, links back to Pasture. But the link is not really discussed or mentioned as often as it is in other quests. This got a little hazy around line 189 when you are creating a profile class using 'pasture' but you have not really discussed it at all prior to this.

Line 253-257: This seems like a topic that needs a conclusion. Does Puppet provide the tools for it to scale? Will the user be able to apply these concepts in order to do so? Where would they learn about this?

The review section seems a little different than earlier ones in that it focuses on the specifics of the quest, but does not really tie this quest to the larger concept of Puppet. Is there a way to do this in a sentence or two?