Edits for application orchestrator
I am adding Kari as a reviewer to this so she can check my work on the orchestrator naming conventions--and anything/everything else, of course.
Line 168: It skipped from Task 1 to Task 6
Line 186-193: I bolded a few terms here that were italic.
Line 217: I would suggest clarifying what you are instructing them to do here to keep consistent with other quest formats and for clarity. Like 219 states "It will look like this..." but I am not sure if "it" is the output of the above command or not.
Line 318-25: As a Puppet newbie/outsider, this section is really dense. Might just be me, but I felt like a quick reminder of the bigger picture from above might help.
Review: In earlier quests, the review had a bigger picture feel to it at the end and connected to other quests. I would suggest adding that here, too.